Split TreeBotHack into multiple files
